fix: amount bug in ReportService

+++ src/main/java/co/inventario/dto/reportDTO/ProductReportDTO.java
@@ -2,5 +2,5 @@ package co.inventario.dto.reportDTO;
 
 import java.time.LocalDateTime;
 
-public record ProductReportDTO(LocalDateTime buyDate, String name, int amount, Double precioUnidad, Double precioTotal) {
+public record ProductReportDTO(LocalDateTime buyDate, String name, Double amount, Double precioUnidad, Double precioTotal) {
 }
